pyproject.toml
[tool.poetry]
name = "dsa-calculator"
version = "0.1.0"
description = ""
authors = ["Max Vieweg <max.vieweg@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = {extras = ["all"], version = "^0.9.0"}
pandas = "^2.1.4"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
